[refactor] split components on side panel page
